squares and triangles are working. CY in parsing is kinda weird

+++ includes/camera.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   camera.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: alex <alex@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: ffarah <ffarah@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/05 15:10:12 by alex              #+#    #+#             */
-/*   Updated: 2021/03/03 13:23:05 by alex             ###   ########.fr       */
+/*   Updated: 2021/03/06 01:12:19 by ffarah           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ typedef struct		s_camera
 	struct s_camera	*prev;
 	t_vector		dir;
 	t_vector		orig;
-	double			fov;
+	float			fov;
 	int				id;
 }					t_camera;
