+++ includes/scene.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   scene.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: alex <alex@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: ffarah <ffarah@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/05 20:29:54 by alex              #+#    #+#             */
-/*   Updated: 2021/03/04 11:19:42 by alex             ###   ########.fr       */
+/*   Updated: 2021/03/06 01:12:22 by ffarah           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ typedef struct		s_canvas
 	int 			height;
 	int				x_pixel;
 	int				y_pixel;
-	double			ratio;
+	float			ratio;
 }					t_canvas;
 
 typedef struct		s_light
@@ -29,7 +29,7 @@ typedef struct		s_light
 	t_vector		orig;
 	t_vector		color;
 	struct s_light	*next;
-	double			intensity;
+	float			intensity;
 	char			type;
 }					t_light;
 
@@ -47,5 +47,5 @@ typedef struct		s_scene
 
 t_scene		*define_scene(void);
 int					blinn_phong(t_intersect *ans, t_scene *scene);
-int 				col_to_int(t_vector *color, t_vector *intens, double coeff);
+int 				col_to_int(t_vector *color, t_vector *intens, float coeff);
 #endif
