+++ srcs/parser/atoi_atof.c
@@ -6,18 +6,18 @@
 /*   By: ffarah <ffarah@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/05 22:02:49 by alex              #+#    #+#             */
-/*   Updated: 2021/02/12 20:03:19 by ffarah           ###   ########.fr       */
+/*   Updated: 2021/03/06 01:12:38 by ffarah           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minirt.h"
 
-double	atof_modified(char **line)
+float	atof_modified(char **line)
 {
 	long long int	parts[2];
 	int				flag;
 	int				power;
-	double			min;
+	float			min;
 	
 	power = 0;
 	flag = 1;
@@ -32,7 +32,7 @@ double	atof_modified(char **line)
 	}
 	parts[0] = atoi_modified(line);
 	if (**line == ',' || !**line || ft_isspace(**line))
-		return ((double)parts[0] * min);
+		return ((float)parts[0] * min);
 	else if (**line != '.')
 		error_throw(-2);
 	*line += 1;
@@ -44,5 +44,5 @@ double	atof_modified(char **line)
 		*line += 1;
 		power += 1;
 	}
-	return (min * ((double)parts[0] + (double)parts[1] / pow(10, power)));
+	return (min * ((float)parts[0] + (float)parts[1] / pow(10, power)));
 }
